refactor: move addProfileTags function to profile support file
